Fix too much scrolling
If you scroll largely (e.g. Hit space key) when the canopy is in the screen,
the amount of scrolling will be more than the viewport height,
and you'll miss some tweets.

I've tried this:
    .ProfileCanopy.is-locked .ProfileCanopy-inner {
        visibility: hidden;
    }

It fixes the above problem, but it reduces the amount of scrolling
on hitting space key because some browsers adjusts the amount of scrolling
of page scroll in consideration of `position: fixed` elements.
(`visibility: hidden` elements are invisible but affects layout.)

Reference:
https://bugzilla.mozilla.org/show_bug.cgi?id=780345
https://developer.mozilla.org/en-US/docs/Web/CSS/visibility